checkpoint_manager and checkpoint_manager rename
